requirements.txt
streamlit==1.52.2
torch==2.9.1
torchvision==0.24.1
numpy==2.3.5
pandas==2.3.3
scikit-learn==1.8.0
matplotlib==3.10.8
seaborn==0.13.2
Pillow==12.0.0